Moved get_config to utils
